+++ Xenity_Engine/Source/engine/network/network.cpp
@@ -116,7 +116,7 @@ std::shared_ptr<Socket> NetworkManager::GetClientSocket()
 	while (true) /* Boucle infinie. Exercice : améliorez ce code. */
 	{
 		int sinsize = sizeof(csin);
-		if ((newClientSocketId = accept(newSocketId, reinterpret_cast<SOCKADDR*>(&csin), &sinsize)) != INVALID_SOCKET)
+		if ((newClientSocketId = static_cast<int>(accept(newSocketId, reinterpret_cast<SOCKADDR*>(&csin), &sinsize))) != INVALID_SOCKET)
 		{
 			Debug::Print("New client connected");
 			break;
@@ -273,7 +273,6 @@ std::shared_ptr<Socket> NetworkManager::CreateSocket(const std::string& address,
 
 	serv_addr.sin_family = AF_INET;
 	serv_addr.sin_port = htons(port);
-
 	if (inet_pton(AF_INET, address.c_str(), &serv_addr.sin_addr) <= 0)
 	{
 		Debug::PrintError("[NetworkManager::CreateSocket] inet_pton error occured");
